+++ failgood/src/test/kotlin/failgood/docs/ClassTestContextTest.kt
@@ -6,9 +6,16 @@ import strikt.api.expectThat
 import strikt.assertions.isEqualTo
 
 @Test
+@Suppress("unused")
 class ClassTestContextTest {
     val context =
         describe("test context defined in a kotlin class") {
+            it("describes behavior") {
+                expectThat("test").isEqualTo("test")
+            }
+        }
+    val context2 =
+        describe("another test context defined in a kotlin class") {
             it("describes behavior") {
                 expectThat("test").isEqualTo("test")
             }